Extract OutputFormatPanel and improve naming consistency
Extract shared output format UI (Select + JSON Schema textarea) into a
reusable OutputFormatPanel component, eliminating duplication between v1
and v2 advanced-options. Rename boolean to hasOutputSchema/shouldDisplayAsJson
for clarity, and pretty-print JSON output in generation view.
